added "AdWare.Win32.BHO.cw" to crash libraries added "Traffic Compressor" to crash libraries some DC++ SVN merge !!! filelist downloads broken (DON' REPORT IT AS A BUG!) !!! sometimes crashes in QueueManager::UserQueue::remove

+++ client/TimerManager.cpp
@@ -55,7 +55,7 @@ uint64_t TimerManager::getTick() {
 		cycles++;
 	}
 	lastTick = tick;
-	return static_cast<uint64_t>(cycles) * (static_cast<uint64_t>(std::numeric_limits<DWORD>::max()) + 1) + tick;
+	return static_cast<uint64_t>(cycles) * (static_cast<uint64_t>(std::numeric_limits<DWORD>::max()) + 1) + static_cast<uint64_t>(tick);
 #else
 	timeval tv2;
 	gettimeofday(&tv2, NULL);
